added tools and example am

+++ simulations/swim/run-sa.sh
@@ -1,5 +1,14 @@
 #!/bin/sh
-{ ../../../simple_am/simple_am localhost; }&
+
+if [ $# -lt 1 ]; then
+	echo "usage: $0 config [run-number(s)|all [ini-file]]"
+	echo example:
+	echo "  "$0 Reactive 0..2
+	exit 1
+fi
+
+
+{ ../../examples/simple_am/simple_am localhost; }&
 AMPID=$!; 
 
 trap 'kill $AMPID;' INT
